Add tests for empty BICs.
This is an additional change to #36, #37.

Also fix a check for required fields in pain.008.002.02. The check was only
executed if "checkAndSanitize" was true. This could lead to an error when a
required field is used but was not provided. It is just checked if the fields
are provided at all. No checks for validity are run without checkAndSanitize
set to true.